fix: misplaced ownership tag
The load balancer security group is the only SG intended to be flagged
as "owned" by the kubernetes cluster, but I noticed that I was seeing
some other groups tagged that way.

It turns out that the tags were being "leaked" into the scope by
AdditionalTags() because the v1infra.Tags type is a map, which means
it's *always* a pointer.

This commit adds a test for the ownership tags and some defensive copies
to try and avoid passing around a mutable scope-level map.