feat: incorporate review changes
Changes made is response to review. Subnet and routetable tables have
been changed so that the '-' isn't stripped out of the az name. Also
the subnet logic has been changed to ensure some of the tests are
done after the subnet data in spec is updated with the matching data
from the existing subnet.