Preserve all VPCSpec fields
When creating/describing the VPC from AWS, we were creating a new
VPCSpec and then using that for the actual VPCSpec in the AWSCluster,
unintentionally wiping out a couple of newly-added fields to VPCSpec
that are user-intent and not information that comes from AWS. This is a
temporary band-aid that ensures we do not lose these fields during
reconciliation. Without this change, the controller sends a patch to set
AvailabilityZoneSelection and AvailabilityZoneUsageLimit to null, but
because of defaulting, the null patches are essentially
rejected/overwritten by the mutating webhook. The result is that the
AWSCluster is reconciled continuously, over and over. This change fixes
that issue.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>