feat: introduce token methods and tag reconcilation and api refactor
The AWSManagedCluster has had most of its fields moved to the
AWSManagedControlPlane so that the reconciliiation logic is
cleaner. This follows the pattern used by CAPZ for the AKS
implementation.

The controller for AWSManagedCluster just reports back the
control plane endpoint, ready and failure domains to CAPI
from the managed control plane.

In the kubeconfig generation we support both aws-iam-authenticator and
also aws cli for generating the auth token. This is available as an
option on the AWSManagedControlPlane with the default being to use
aws-iam-authenticator.

Tags reconcilation has been added so that if the only things that changes
is the tags then these changes will be realised in the EKS related
resources.

The API group name has been changed to remove the `exp` prefix.