hack: fix semver comparison in ensure-kind
`ensure-kind.sh` should compare semantic versions correctly, instead of
checking for the exact number.
Otherwise it could fail even when the local kind version is v0.10.0,
which is obviously greater than the minimum version v0.9.0.

So let's follow mostly the way how other ensure scripts are implemented:
read the kind version, run version sort with the minimum version,
and compare it to the minimum version.