chore: Remove unnecessary conversion code
The removed code, introduced in #2556, became unnecessary when
equivalent code was added to the restoreInstance() function in #2468.

This code handles a subset of the cases that restoreInstances() does;
namely, when both the restored and dst pointers are non-nil. The
conversion tests fail if the restoreInstance() call is removed, but this
code is left.  They pass if this code is removed.

Signed-off-by: Daniel Lipovetsky <dlipovetsky@d2iq.com>